Flatten BinaryNbtReader/Writer abstraction
There is now a new Java network variant that changes the actual layout of the serialized data (making root tags nameless). So, since a variant can vary by more than just the way primitives are encoded, it makes sense to eliminate the NBT binary sinks and move that into the reader/writer.