Add `NbtVariant.JavaNetwork` to support protocol version 764
This new protocol version encodes NBT tags without a root name at all, just a tag type then the payload.

All protocol versions will treat the tag as unnamed, with the old protocol versions implicitly having an empty named tag. This allows the in-memory representation of the data to be the same between the old and the new protocols. The root name being an empty string should be fine since all packets with the old format used empty string root names.