Fix decoding from Java `InputStream` reading past the end of the tag
With the okio API this happens too, but since the caller is passing in the `BufferedSource`, they still have access to the unused bytes that are read. For the `InputStream` though, the bytes are sitting in a `BufferedSource`, though the caller doesn't have access to the source or its bytes.