Change knbt license to LGPL
Originally the GPL 3.0 license was decided on so that any improvements or derivative projects that benefit from this project would also have to their source be shared and open.

However, GPL imposes itself even if this library is only being used as a dependency. LGPL would've been more in line with the original intent for the license, since it more so applies to modification of the project/binaries, and doesn't impose itself on other projects when used as a dependency.