Support built-in polymorphic serializers
This was implemented as an adapting serializer that intercepts the calls from the `AbstractPolymorphicSerializer` (which the builtin polymorphic/sealed serializers use) to retreive the polymorphic type/value. Then, the serialization simply works with the `NbtCompound` tag representation to deal with the class discriminator.

Implementing it this way likely isn't as performant, since an intermediate `NbtTag` structure is instanttiated, but this way is a lot less obtrusive, and is more of a stopgap until the next big release where the polymorphic serializer support could be implemented more closely to the NBT encoder/decoder.

This polymorphic serializer support also follows the `SerialName`-nested root tag behavior that classes have, as the data it serializes is declared/represented the same way, as a class/interface that's automatically serialized as a NbtCompound, just additionally marked as sealed/polymorphic.