Update Kotlin to 1.9.23
Update Okio, since it depended on the legacy metadata and JS backend. Changing major versions from Okio 2.9.0 -> 3.0.0 should not be an issue for compatibility according to the Okio changelog:
> This is the first stable release of Okio 3.x. This release is strongly backwards-compatible with Okio 2.x, and the new major version signifies new capabilities more than it does backwards incompatibility.

Kotlin's hierarchical multiplatform means symbolic links no longer need to be used to share the native source set. Now having the nativeMain directory in src is enough.

OptIn has now been stabilized, and because of changes to contagiousness, many of the `@OptIn` markers are no longer needed.

Objects with their class name for `toString()` were also changed to be data objects, now that they're available in Kotlin.

It seems that Kotlin/IR has an issue with `NbtException` accessing `super.message`, causing infinite recursion when called from `this.message` when the legacy JS backend didn't have a problem. This was worked around by storing the super class's `message` in a private property and using that instead.