Add `Parameter.reportedAs()` for parameterized tests
Not all arguments have a good `toString()` when listed in parameterized
test failures. Or sometimes the test depends on one part of the
argument, so it's more useful to report that part with its name.

This function changes how the parameter is declared so that a more
useful parameter/argument list can be shown when the test fails.