Change `NbtFormat` into a class
Internal API will be added to `NbtFormat` for the encoder/decoder to
use, and that's not possible with interfaces yet in Kotlin.

Additionally, inheritance does make sense here since its encode/decode
functions truely are being inherited, with the stringified/binary
implementations extending its behavior.

There also is a possibility of this class being usable on its own in the
future, with its own builder so serialization to/from `NbtTag`s can be
done without any extra baggage from the other variants of NBT.