WIP Move named root check into encoder/decoder
nbtwriter/reader should be dumb, only validating the values read, and
leaving any other validation to the encoder/decoder

TODO ^document in writer/reader interfaces

capabilities part of `NbtFormat` instead of the `NbtWriter`/`NbtReader`
so that serializing to `NbtTag` has the same restrictions as the
format-specific encode/decode functions. This is useful for
consistency, knowing that the actual serialization will fail if the
NbtTag serialization does too, and allowing serialization issues to be
caught in tests.

TODO ^docs in `encodeTo`/`decodeFromNbtTag`

Removed check in BinaryNbtReader since it was already guarded in the
decoder