Remove path tracking from `NbtReaderDecoder`
There are problems with the way it was implemented, such as
`getPathNode()` accessing uninitialized lateinit vars after failures
during construction. There's also issues with the way it's designed,
assuming that each of the decoder's CompositeDecoder implementations is
one node in the NBT path, when in reality they can represent one of its
element's nodes, or the node for structure itself when between elements
(with potentially incomplete information if not decoded yet).

Path tracking will be implemented more efficiently and robustly in the
future using the `NbtContext` to share path logic between the encoder
and the decoder.